Build demos for each nav example
