+++ src/patternfly/demos/PageLayout/examples/page-layout-default-nav-example.hbs
@@ -1,0 +1,58 @@
+{{#> background-image}}
+{{/background-image}}
+
+{{#> page page--attribute='id="page"'}}
+  {{#> page-header}}
+    {{!-- Brand --}}
+    {{#> page-header-brand}}
+      {{#> page-header-brand-group}}
+        {{#> nav-toggle nav-toggle--attribute='id="nav-toggle"'}}
+          <i class="fas fa-bars"></i>
+        {{/nav-toggle}}
+        <a class="pf-l-page__header-brand-link" href="#">
+          <img class="pf-l-page__header-brand-logo" src="/assets/images/l_pf-reverse.svg" alt="Patternfly brand logo" alt="Patternfly Logo" /> 
+        </a>
+      {{/page-header-brand-group}}  
+    {{/page-header-brand}}
+    {{#> page-template-header-elements}}
+    {{/page-template-header-elements}}
+  {{/page-header}}
+
+  {{!-- Nav --}}
+  {{#> nav nav--attribute='aria-label="Primary Nav Default Example"'}}
+    {{#> nav-list}}
+      {{#> nav-list-item}}
+        {{#> nav-list-link nav-list-link--href="#" nav-list-link--current="true"}}
+          System Panel
+        {{/nav-list-link}}
+      {{/nav-list-item}}
+      {{#> nav-list-item}}
+        {{#> nav-list-link nav-list-link--href="#"}}
+          Policy
+        {{/nav-list-link}}
+      {{/nav-list-item}}
+      {{#> nav-list-item}}
+        {{#> nav-list-link nav-list-link--href="#"}}
+          Authentication
+        {{/nav-list-link}}
+      {{/nav-list-item}}
+      {{#> nav-list-item}}
+        {{#> nav-list-link nav-list-link--href="#"}}
+          Network Services
+        {{/nav-list-link}}
+      {{/nav-list-item}}
+      {{#> nav-list-item}}
+        {{#> nav-list-link nav-list-link--href="#"}}
+          Server
+        {{/nav-list-link}}
+      {{/nav-list-item}}
+    {{/nav-list}}
+  {{/nav}}
+
+  {{#> page-main}}
+    {{#> page-template-title}}
+    {{/page-template-title}}
+    {{#> page-template-gallery}}
+    {{/page-template-gallery}}
+  {{/page-main}}
+{{/page}}
