+++ src/patternfly/demos/PageLayout/examples/page-layout-expandable-nav-example.hbs
@@ -1,0 +1,111 @@
+{{#> background-image}}
+{{/background-image}}
+
+{{#> page page--attribute='id="page"'}}
+  {{#> page-header}}
+
+    {{!-- Brand --}}
+    {{#> page-header-brand}}
+      {{#> page-header-brand-group}}
+        {{#> nav-toggle nav-toggle--attribute='id="nav-toggle"'}}
+          <i class="fas fa-bars"></i>
+        {{/nav-toggle}}
+        <a class="pf-l-page__header-brand-link" href="#">
+          <img class="pf-l-page__header-brand-logo" src="/assets/images/l_pf-reverse.svg" alt="Patternfly brand logo" alt="Patternfly Logo" /> 
+        </a>
+      {{/page-header-brand-group}}  
+    {{/page-header-brand}}
+    {{#> page-template-header-elements}}
+    {{/page-template-header-elements}}
+  {{/page-header}}
+
+  {{#> nav nav--attribute='aria-label="Primary Nav Expandable Example"'}}
+    {{#> nav-list nav-list--attribute='id="primary-nav"'}}
+      {{#> nav-list-item nav-list-item--expandable="true" nav-list-item--expanded="true" nav-list-item--current="true"}}
+        {{#> nav-list-link nav-list-link--href="#" nav-list-link--attribute='id="link1"'}}
+          System Panel
+        {{/nav-list-link}}
+        {{#> nav-subnav nav-subnav--attribute='aria-labelledby="link1"'}}
+          {{#> nav-list-simple}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+                Overview
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#" nav-list-simple-link--current="true"}}
+                Resource Usage
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+                Hypervisors
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+                Instances
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+                Volumes
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+                Networks
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+          {{/nav-list-simple}}
+        {{/nav-subnav}}
+      {{/nav-list-item}}
+      {{#> nav-list-item nav-list-item--expandable="true"}}
+        {{#> nav-list-link nav-list-link--href="#" nav-list-link--attribute='id="link2"'}}
+          Policy
+        {{/nav-list-link}}
+        {{#> nav-subnav nav-subnav--attribute='aria-labelledby="link2"'}}
+          {{#> nav-list-simple}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+                Subnav link 1
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+                Subnav link 2
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+          {{/nav-list-simple}}
+        {{/nav-subnav}}
+      {{/nav-list-item}}
+      {{#> nav-list-item nav-list-item--expandable="true"}}
+        {{#> nav-list-link nav-list-link--href="#" nav-list-link--attribute='id="link3"'}}
+          Authentication
+        {{/nav-list-link}}
+        {{#> nav-subnav nav-subnav--attribute='aria-labelledby="link3"'}}
+          {{#> nav-list-simple}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+                Subnav link 1
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+            {{#> nav-list-simple-item}}
+              {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+                Subnav link 2
+              {{/nav-list-simple-link}}
+            {{/nav-list-simple-item}}
+          {{/nav-list-simple}}
+        {{/nav-subnav}}
+      {{/nav-list-item}}
+    {{/nav-list}}
+  {{/nav}}
+  {{#> page-main}}
+    {{#> page-template-tertiary-nav}}
+    {{/page-template-tertiary-nav}}
+    {{#> page-template-title}}
+    {{/page-template-title}}
+    {{#> page-template-gallery}}
+    {{/page-template-gallery}}
+  {{/page-main}}
+{{/page}}
