+++ src/patternfly/demos/PageLayout/examples/page-layout-grouped-nav-example.hbs
@@ -1,0 +1,100 @@
+{{#> background-image}}
+{{/background-image}}
+
+{{#> page page--attribute='id="page"'}}
+  {{#> page-header}}
+
+    {{!-- Brand --}}
+    {{#> page-header-brand}}
+      {{#> page-header-brand-group}}
+        {{#> nav-toggle nav-toggle--attribute='id="nav-toggle"'}}
+          <i class="fas fa-bars"></i>
+        {{/nav-toggle}}
+        <a class="pf-l-page__header-brand-link" href="#">
+          <img class="pf-l-page__header-brand-logo" src="/assets/images/l_pf-reverse.svg" alt="Patternfly brand logo" alt="Patternfly Logo" /> 
+        </a>
+      {{/page-header-brand-group}}  
+    {{/page-header-brand}}
+    {{#> page-header-selector}}
+      pf-c-context-selector
+    {{/page-header-selector}}
+    {{#> page-template-header-elements}}
+    {{/page-template-header-elements}}
+  {{/page-header}}
+
+  {{#> nav nav--attribute='aria-label="Primary Nav Grouped Example"'}}
+    {{#> nav-section nav-section--attribute='aria-labelledby="grouped-title1"'}}
+      {{#> nav-section-title nav-section-title--attribute='id="grouped-title1"'}}
+        System Panel
+      {{/nav-section-title}}
+      {{#> nav-list-simple}}
+        {{#> nav-list-simple-item}}
+          {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+            Overview
+          {{/nav-list-simple-link}}
+        {{/nav-list-simple-item}}
+        {{#> nav-list-simple-item}}
+          {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+            Resource Usage
+          {{/nav-list-simple-link}}
+        {{/nav-list-simple-item}}
+        {{#> nav-list-simple-item}}
+          {{#> nav-list-simple-link nav-list-simple-link--href="#" nav-list-simple-link--current="true"}}
+            Hypervisors
+          {{/nav-list-simple-link}}
+        {{/nav-list-simple-item}}
+        {{#> nav-list-simple-item}}
+          {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+            Instances
+          {{/nav-list-simple-link}}
+        {{/nav-list-simple-item}}
+        {{#> nav-list-simple-item}}
+          {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+            Volumes
+          {{/nav-list-simple-link}}
+        {{/nav-list-simple-item}}
+        {{#> nav-list-simple-item}}
+          {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+            Networks
+          {{/nav-list-simple-link}}
+        {{/nav-list-simple-item}}
+      {{/nav-list-simple}}
+    {{/nav-section}}
+    {{#> nav-section nav-section--attribute='aria-labelledby="grouped-title2"'}}
+      {{#> nav-section-title nav-section-title--attribute='id="grouped-title2"'}}
+        Policy
+      {{/nav-section-title}}
+      {{#> nav-list-simple}}
+        {{#> nav-list-simple-item}}
+          {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+            Hosts
+          {{/nav-list-simple-link}}
+        {{/nav-list-simple-item}}
+        {{#> nav-list-simple-item}}
+          {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+            Virtual Machines
+          {{/nav-list-simple-link}}
+        {{/nav-list-simple-item}}
+        {{#> nav-list-simple-item}}
+          {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+            Storage
+          {{/nav-list-simple-link}}
+        {{/nav-list-simple-item}}
+      {{/nav-list-simple}}
+    {{/nav-section}}
+  {{/nav}}
+
+  {{#> page-main}}
+    {{#> page-template-title}}
+    {{/page-template-title}}
+    {{#> page-main-section page-main-section--modifier="pf-m-light"}}
+      Light content
+    {{/page-main-section}}
+    {{#> page-main-section page-main-section--modifier="pf-m-dark"}}
+      Dark content
+    {{/page-main-section}}
+    {{#> page-main-section}}
+      Content
+    {{/page-main-section}}
+  {{/page-main}}
+{{/page}}
