+++ src/patternfly/demos/PageLayout/examples/page-layout-horizontal-nav-example.hbs
@@ -81,111 +81,11 @@
   {{#> page-main}}
 {{!--     {{#> page-main-section page-main-section--modifier="pf-m-dark"}}
       Dark content
-    {{/page-main-section}}
-
-    {{#> page-main-section page-main-section--modifier="pf-m-light"}}
-      Light content
     {{/page-main-section}} --}}
-
-    {{#> page-main-section}}
-      {{#> gallery modifierClass="pf-m-gutter"}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-      {{/gallery}}
-    {{/page-main-section}}
+    {{#> page-template-title}}
+    {{/page-template-title}}
+    {{#> page-template-gallery}}
+    {{/page-template-gallery}}
   {{/page-main}}
 {{/page}}
 
-<script type="text/javascript">
-  const navToggle = document.getElementById("nav-toggle");
-  document.getElementById("nav-toggle").addEventListener("click", classToggle);
-</script>
