+++ src/patternfly/demos/PageLayout/examples/page-layout-simple-nav-example.hbs
@@ -1,0 +1,62 @@
+{{#> background-image}}
+{{/background-image}}
+
+{{#> page page--attribute='id="page"'}}
+  {{#> page-header}}
+
+    {{!-- Brand --}}
+    {{#> page-header-brand}}
+      {{#> page-header-brand-group}}
+        {{#> nav-toggle nav-toggle--attribute='id="nav-toggle"'}}
+          <i class="fas fa-bars"></i>
+        {{/nav-toggle}}
+        <a class="pf-l-page__header-brand-link" href="#">
+          <img class="pf-l-page__header-brand-logo" src="/assets/images/l_pf-reverse.svg" alt="Patternfly brand logo" alt="Patternfly Logo" /> 
+        </a>
+      {{/page-header-brand-group}}  
+    {{/page-header-brand}}
+    {{#> page-template-header-elements}}
+    {{/page-template-header-elements}}
+  {{/page-header}}
+
+  {{#> nav nav--attribute='id="nav-primary-simple" aria-label="Primary Nav, Simple List Example"'}}
+    {{#> nav-list-simple}}
+      {{#> nav-list-simple-item}}
+        {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+          Overview
+        {{/nav-list-simple-link}}
+      {{/nav-list-simple-item}}
+      {{#> nav-list-simple-item}}
+        {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+          Resource Usage
+        {{/nav-list-simple-link}}
+      {{/nav-list-simple-item}}
+      {{#> nav-list-simple-item}}
+        {{#> nav-list-simple-link nav-list-simple-link--href="#" nav-list-simple-link--current="true"}}
+          Hypervisors
+        {{/nav-list-simple-link}}
+      {{/nav-list-simple-item}}
+      {{#> nav-list-simple-item}}
+        {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+          Instances
+        {{/nav-list-simple-link}}
+      {{/nav-list-simple-item}}
+      {{#> nav-list-simple-item}}
+        {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+          Volumes
+        {{/nav-list-simple-link}}
+      {{/nav-list-simple-item}}
+      {{#> nav-list-simple-item}}
+        {{#> nav-list-simple-link nav-list-simple-link--href="#"}}
+          Networks
+        {{/nav-list-simple-link}}
+      {{/nav-list-simple-item}}
+    {{/nav-list-simple}}
+  {{/nav}}
+  {{#> page-main}}
+    {{#> page-template-title}}
+    {{/page-template-title}}
+    {{#> page-template-gallery}}
+    {{/page-template-gallery}}
+  {{/page-main}}
+{{/page}}
