+++ src/patternfly/demos/PageLayout/page-template-gallery.hbs
@@ -1,0 +1,74 @@
+{{#> page-main-section}}
+  {{#> gallery modifierClass="pf-m-gutter"}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+    {{#> gallery-item}}
+      {{#> card }}
+        {{#> card-body }}
+          This is a card
+        {{/card-body}}
+      {{/card}}
+    {{/gallery-item}}
+  {{/gallery}}
+{{/page-main-section}}
