+++ src/patternfly/demos/PageLayout/page-template-tertiary-nav.hbs
@@ -1,0 +1,29 @@
+{{#> nav nav--modifier="pf-m-nav-tertiary" nav--attribute='aria-label="Tertiary Example"' aria-labelledby="[id of parent]"}}
+  {{#> nav-list nav-list--type="tertiary"}}
+    {{#> nav-list-item}}
+      {{#> nav-list-link nav-list-link--href="#" nav-list-link--current="true"}}
+        Nav Item 1
+      {{/nav-list-link}}
+    {{/nav-list-item}}
+    {{#> nav-list-item}}
+      {{#> nav-list-link nav-list-link--href="#"}}
+        Nav Item 2
+      {{/nav-list-link}}
+    {{/nav-list-item}}
+    {{#> nav-list-item}}
+      {{#> nav-list-link nav-list-link--href="#"}}
+        Nav Item 3
+      {{/nav-list-link}}
+    {{/nav-list-item}}
+    {{#> nav-list-item}}
+      {{#> nav-list-link nav-list-link--href="#"}}
+        Nav Item 4
+      {{/nav-list-link}}
+    {{/nav-list-item}}
+    {{#> nav-list-item}}
+      {{#> nav-list-link nav-list-link--href="#"}}
+        Nav Item 5
+      {{/nav-list-link}}
+    {{/nav-list-item}}
+  {{/nav-list}}
+{{/nav}}
