+++ src/patternfly/demos/PageLayout/page-template-title.hbs
@@ -1,0 +1,11 @@
+{{#> page-main-section page-main-section--modifier="pf-m-light"}}
+  {{#> content}}
+    {{#> title titleType="h1" pf-c-title--Modifier="pf-m-3xl"}}
+      Main title
+    {{/title}}
+    <p>
+      Body text should be Overpass Regular at 16px. It should have leading of 24px because <br>
+      of it’s relative line height of 1.5.
+    </p>
+  {{/content}}
+{{/page-main-section}}
