+++ src/patternfly/demos/PageLayout/page-template.hbs
@@ -1,5 +1,7 @@
 {{#> background-image}}
 {{/background-image}}
+{{#> page page--nav-vertical="true" page--attribute='id="page"'}}
+
 
 {{#> page page--nav-vertical="true" page--attribute='id="page"'}}
   {{#> page-header}}
@@ -188,184 +190,8 @@
       {{/content}}
     {{/page-main-section}}
     {{#> page-main-section}}
-      {{#> gallery modifierClass="pf-m-gutter"}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-          {{#> card }}
-            {{#> card-body }}
-              This is a card
-            {{/card-body}}
-          {{/card}}
-        {{/gallery-item}}
-
-      {{/gallery}}
+      {{#> page-template-gallery}}
+      {{/page-template-gallery}}
     {{/page-main-section}}
   {{/page-main}}
 {{/page}}
