--- src/patternfly/demos/PageTest/examples/page-test-example3.hbs
@@ -1,166 +1,0 @@
-{{#> background-image}}
-{{/background-image}}
-
-{{#> page page--nav-horizontal="true" page--attribute='id="page"'}}
-  {{#> page-header}}
-
-    {{!-- Brand --}}
-    {{#> page-header-brand page-header-brand--modifier="pf-u-sr-only pf-u-visible-on-md"}}
-      {{#> nav-toggle nav-toggle--attribute='id="nav-toggle"'}}
-        <i class="fas fa-bars"></i>
-      {{/nav-toggle}}
-      <a class="pf-l-page__header-brand-link" href="#">
-        <img class="pf-l-page__header-brand-logo pf-u-sr-only pf-u-visible-on-md" src="/assets/images/l_pf-reverse.svg" alt="Patternfly brand logo" alt="Patternfly Logo" /> 
-      </a>
-    {{/page-header-brand}}
-
-    {{!-- Context Selector --}}
-    {{#> page-header-selector}}
-      pf-c-context-selector
-    {{/page-header-selector}}
-
-    {{!-- Horizontal Nav --}}
-    {{#> page-header-nav}}
-      {{#> nav nav--attribute='id="nav-primary" aria-label="Primary Nav Default Example"'}}
-        {{#> nav-list nav-list--type="horizontal"}}
-          {{#> nav-list-item}}
-            {{#> nav-list-link nav-list-link--href="#" nav-list-link--current="true"}}
-              Current link
-            {{/nav-list-link}}
-          {{/nav-list-item}}
-          {{#> nav-list-item}}
-            {{#> nav-list-link nav-list-link--href="#"}}
-              Link 2
-            {{/nav-list-link}}
-          {{/nav-list-item}}
-          {{#> nav-list-item}}
-            {{#> nav-list-link nav-list-link--href="#"}}
-              Link 3
-            {{/nav-list-link}}
-          {{/nav-list-item}}
-          {{#> nav-list-item}}
-            {{#> nav-list-link nav-list-link--href="#"}}
-              Link 4
-            {{/nav-list-link}}
-          {{/nav-list-item}}
-          {{#> nav-list-item}}
-            {{#> nav-list-link nav-list-link--href="#"}}
-              Link 5
-            {{/nav-list-link}}
-          {{/nav-list-item}}
-        {{/nav-list}}
-      {{/nav}}
-    {{/page-header-nav}}
-
-    {{!-- Toolbar --}}
-    {{#> page-header-tools}}
-      pf-l-toolbar
-    {{/page-header-tools}}
-  {{/page-header}}
-
-  {{#> page-main}}
-{{!--     {{#> page-main-section page-main-section--modifier="pf-m-dark"}}
-      Dark content
-    {{/page-main-section}}
-
-    {{#> page-main-section page-main-section--modifier="pf-m-light"}}
-      Light content
-    {{/page-main-section}} --}}
-
-    {{#> page-main-section}}
-      {{#> gallery modifierClass="pf-m-gutter"}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-        {{#> gallery-item}}
-        {{#> card }}
-          {{#> card-body }}
-            This is a card
-          {{/card-body}}
-        {{/card}}
-        {{/gallery-item}}
-      {{/gallery}}
-    {{/page-main-section}}
-  {{/page-main}}
-{{/page}}
