--- src/patternfly/demos/PageTest/page-test.hbs
@@ -1,6 +1,0 @@
-<div class="pf-d-page-test {{pf-d-page-test--modifier}}"
-  {{#if pf-d-page-test__id}}
-    id="{{pf-d-page-test__id}}"
-  {{/if}}>
-  {{> @partial-block}}
-</div>
